fixCartPage
