+++ src/pages/CartPage/CartPage.js
@@ -29,6 +29,8 @@ export default function CartPage() {
   const [isChecked, setIsChecked] = useState(false);
   const [isShow, setIsShow] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
+  const [pickupTime, setPickupTime] = useState("");
+  const [remarks, setRemarks] = useState("");
   const cartData = useSelector(selectCartData);
   const userId = useSelector(selectUserId);
   const cart = useSelector(selectCart);
@@ -80,6 +82,8 @@ export default function CartPage() {
   const handleCheckedClick = (e) => {
     dispatch(setVendorId(e.target.value));
     setIsChecked(!isChecked);
+    setPickupTime("");
+    setRemarks("");
   };
 
   const handleDeleteClick = (id, userId) => {
@@ -90,7 +94,14 @@ export default function CartPage() {
     dispatch(setCartData(newCartData));
   };
 
-  const handleSubmit = (orderProducts, vendorId, pickupTime, remarks) => {
+  const handleSubmit = (
+    orderProducts,
+    vendorId,
+    pickupTime,
+    remarks,
+    userId,
+    cartData
+  ) => {
     if (!pickupTime) {
       setIsShow(false);
       dispatch(setErrorMessage("請填寫預約時間!"));
@@ -103,19 +114,14 @@ export default function CartPage() {
         vendorId,
         pickupTime,
         remarks,
+        userId,
+        cartData,
       })
     );
     setIsShow(false);
-    let newData = [].concat(
-      JSON.parse(cartData).filter((obj1) =>
-        orderProducts.every((obj2) => obj1.id !== obj2.id)
-      ),
-      orderProducts.filter((obj2) =>
-        JSON.parse(cartData).every((obj1) => obj2.id !== obj1.id)
-      )
-    );
-    localStorage.setItem(`cartId${userId}`, JSON.stringify(newData));
-    dispatch(setCartData(JSON.stringify(newData)));
+    setIsChecked(false);
+    setPickupTime("");
+    setRemarks("");
   };
 
   return (
@@ -198,6 +204,10 @@ export default function CartPage() {
             handleIsShow={handleIsShow}
             handleSubmit={handleSubmit}
             userId={userId}
+            pickupTime={pickupTime}
+            remarks={remarks}
+            setPickupTime={setPickupTime}
+            setRemarks={setRemarks}
           />
         </>
       )}
